+++ src/types.d.ts
@@ -40,6 +40,75 @@ export type S3Payload = {
   objectUrl: Scalars["String"]
 }
 
+export type Discussion = Base & {
+  __typename?: "Discussion"
+  id: Scalars["ID"]
+  PK: Scalars["String"]
+  SK: Scalars["String"]
+  created: Scalars["Date"]
+  updated?: Maybe<Scalars["Date"]>
+  title: Scalars["String"]
+  content: Scalars["String"]
+  authorId: Scalars["String"]
+  comments?: Maybe<Array<Maybe<Comment>>>
+}
+
+export type Comment = Base & {
+  __typename?: "Comment"
+  id: Scalars["ID"]
+  PK: Scalars["String"]
+  SK: Scalars["String"]
+  created: Scalars["Date"]
+  updated?: Maybe<Scalars["Date"]>
+  content: Scalars["String"]
+  authorId: Scalars["String"]
+  replyToId?: Maybe<Scalars["String"]>
+  replies?: Maybe<Array<Maybe<Comment>>>
+}
+
+export type CommentReaction = Base & {
+  __typename?: "CommentReaction"
+  id: Scalars["ID"]
+  PK: Scalars["String"]
+  SK: Scalars["String"]
+  created: Scalars["Date"]
+  updated?: Maybe<Scalars["Date"]>
+  reaction: Scalars["String"]
+}
+
+export type GetDiscussionsKey = {
+  __typename?: "GetDiscussionsKey"
+  PK?: Maybe<Scalars["String"]>
+  SK?: Maybe<Scalars["String"]>
+  created?: Maybe<Scalars["Date"]>
+}
+
+export type LastEvaluatedKey = {
+  PK?: Maybe<Scalars["String"]>
+  SK?: Maybe<Scalars["String"]>
+  created?: Maybe<Scalars["Date"]>
+}
+
+export type GetDiscussionsQueryResult = {
+  __typename?: "GetDiscussionsQueryResult"
+  items?: Maybe<Array<Maybe<Discussion>>>
+  lastEvaluatedKey?: Maybe<GetDiscussionsKey>
+}
+
+export type CreateDiscussionInput = {
+  title: Scalars["String"]
+  content: Scalars["String"]
+  authorId: Scalars["String"]
+}
+
+export type CreateCommentInput = {
+  content: Scalars["String"]
+  discussionId: Scalars["String"]
+  /** The id of another comment */
+  authorId: Scalars["String"]
+  replyToId?: Maybe<Scalars["String"]>
+}
+
 /**
  * The typical response shape from AWS Cognito
  * - expect this from the /oauth2/token endpoint
@@ -106,10 +175,20 @@ export type User = Base & {
 
 export type Query = {
   __typename?: "Query"
+  getDiscussions?: Maybe<GetDiscussionsQueryResult>
+  getDiscussionById?: Maybe<Discussion>
   /** 🔒 This field requires you to be authenticated */
   getOrCreateUser: User
   /** 🔒 This field requires you to be authenticated */
   getUsers?: Maybe<Array<Maybe<User>>>
+}
+
+export type QueryGetDiscussionsArgs = {
+  lastEvaluatedKey?: Maybe<LastEvaluatedKey>
+}
+
+export type QueryGetDiscussionByIdArgs = {
+  id: Scalars["ID"]
 }
 
 export type QueryGetOrCreateUserArgs = {
@@ -124,6 +203,8 @@ export type Mutation = {
   __typename?: "Mutation"
   /** 🔒 This field requires you to be authenticated */
   s3GetSignedPutObjectUrl: S3Payload
+  createDiscussion?: Maybe<Discussion>
+  createComment?: Maybe<Comment>
   /** Trade a code—appended by the Cognito Hosted UI—for Cognito Tokens */
   getToken?: Maybe<AuthResponse>
   /** 🔒 This field requires you to be authenticated */
@@ -135,6 +216,14 @@ export type Mutation = {
 export type MutationS3GetSignedPutObjectUrlArgs = {
   fileName: Scalars["String"]
   fileType: Scalars["String"]
+}
+
+export type MutationCreateDiscussionArgs = {
+  input: CreateDiscussionInput
+}
+
+export type MutationCreateCommentArgs = {
+  input: CreateCommentInput
 }
 
 export type MutationGetTokenArgs = {