+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v0.7.0] - 2020-12-14
 
+This version encompasses work for a POC of recreating GitHub's 'Discussions' feature
+
 ### Added
 
 - installed `yup` to validate requests that pass GraphQL schema validation, but don't actually have valid values
+- `Query.getDiscussionById`
+- `Query.getDiscussions`
+- `Mutation.createDiscussion`
+- `Mutation.createComment`
 
 #### DynamoDB `GSI`'s
 
@@ -17,6 +23,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Partition: `'SK'`; Sort: `'created'`
 - **`GSI_Root_Created`**
   - Partition: `'Root'`; Sort: `'created'`
+
+### Updated
+
+- types.d.ts
+- schema
 
 ## [v0.6.3] - 2020-12-10